Refactored imports to be consistent.
I moved printing functions/classes into the printing module. Moved the
VectorTypeError into vector. Set mechanics_printing to equal the new
vector.init_printing. Updated doctests to reflect the new behavior. Removed the
time_derivative_printing function in favor of init_printing (no dep warnings
needed, as it was never released, same goes for a lot of this).
