+++ sympy/physics/mechanics/__init__.py
@@ -31,36 +31,17 @@ from . import lagrange
 from .lagrange import *
 __all__.extend(lagrange.__all__)
 
+from sympy.physics import vector
+from sympy.physics.vector import *
+__all__.extend(vector.__all__)
 
-#Import essential elements from physics.vector module
-from sympy.physics.vector.frame import ReferenceFrame, CoordinateSym
-from sympy.physics.vector.dyadic import Dyadic
-from sympy.physics.vector.vector import Vector
-from sympy.physics.vector.printing import (
-    VectorStrPrinter as MechanicsStrPrinter,
-    VectorLatexPrinter as MechanicsLatexPrinter,
-    VectorPrettyPrinter as MechanicsPrettyPrinter)
-from sympy.physics.vector.point import Point
-from sympy.physics.vector.functions import (cross, dot, express,
-                                            time_derivative, outer,
-                                            kinematic_equations,
-                                            get_motion_params,
-                                            partial_velocity,
-                                            dynamicsymbols)
-from sympy.physics.vector.functions import (
-     time_derivative_printing as mechanics_printing,
-     vprint as mprint, vsprint as msprint,
-     vpprint as mpprint, vlatex as mlatex)
+# These are functions that we've renamed in during the extraction of the
+# basic vector calculus code from the mechanics packages.
+# TODO : Add deprecation warnings for using these:
 
-#essentialnames contains all names to be imported from vector package
-essentialnames = ['ReferenceFrame', 'CoordinateSym',
-                  'Dyadic', 'Vector', 'MechanicsStrPrinter',
-                  'MechanicsLatexPrinter',
-                  'MechanicsPrettyPrinter', 'dynamicsymbols',
-                  'Point', 'cross', 'dot', 'express',
-                  'time_derivative', 'outer', 'kinematic_equations',
-                  'get_motion_params', 'partial_velocity',
-                  'mechanics_printing', 'mprint', 'msprint',
-                  'mpprint', 'mlatex']
+mechanics_printing = init_printing
+msprint = vsprint
+mpprint = vpprint
+mlatex = vlatex
 
-__all__.extend(essentialnames)
+__all__.extend(['mechanics_printing', 'msprint', 'mpprint', 'mlatex'])
