+++ sympy/physics/vector/tests/test_frame.py
@@ -1,6 +1,6 @@
 from sympy import sin, cos, pi, zeros, ImmutableMatrix as Matrix
-from sympy.physics.vector import ReferenceFrame, Vector, CoordinateSym, \
-     dynamicsymbols, time_derivative, express
+from sympy.physics.vector import (ReferenceFrame, Vector, CoordinateSym,
+                                  dynamicsymbols, time_derivative, express)
 
 
 Vector.simp = True
