+++ sympy/physics/vector/tests/test_output.py
@@ -1,8 +1,6 @@
-from sympy import S, Function
-from sympy.physics.vector import Vector, ReferenceFrame, Dyadic, \
-     VectorLatexPrinter
+from sympy import S
+from sympy.physics.vector import Vector, ReferenceFrame, Dyadic
 from sympy.utilities.pytest import raises
-
 
 Vector.simp = True
 A = ReferenceFrame('A')
